feat: dropdown 바깥쪽 클릭 시 닫힘 (#77)
