+++ src/components/dashboard/side/Menus.tsx
@@ -1,4 +1,4 @@
-import ActiveLink from './ActiveLink';
+import ActiveLink from '../ActiveLink';
 
 export default function Menus() {
   return (
