+++ src/pages/contact/unconfirmed/index.tsx
@@ -1,8 +1,8 @@
 import Header from '@/components/common/Header';
 import ContactContents from '@/components/dashboard/contact/ContactContents';
-import Sidebar from '@/components/dashboard/Sidebar';
+import Sidebar from '@/components/dashboard/side/Sidebar';
 
-export default function ContactConfirmNeeded() {
+export default function Unconfirmed() {
   return (
     <div>
       <Header />
